fix(ngrx-notification): compilation error
Due to upgrading to Angular 16,  `REDUCERS` were in the incorrect file and
should have been moved to `app.config.ts` from `main.ts`.
