+++ apps/ngrx-notification/src/app/app.config.ts
@@ -5,9 +5,23 @@ import { TeacherEffects } from './teacher/store/teacher.effects';
 import { StudentEffects } from './student/store/student.effects';
 import { provideRouter } from '@angular/router';
 import { ROUTES } from './routes';
-import { APP_INITIALIZER, enableProdMode, inject } from '@angular/core';
+import { APP_INITIALIZER, inject } from '@angular/core';
 import { FakeBackendService } from '@angular-challenges/ngrx-notification/backend';
 import { NotificationService } from './data-access/notification.service';
+import {
+  teacherReducer,
+  teachersFeatureKey,
+} from './teacher/store/teacher.reducer';
+import {
+  studentReducer,
+  studentsFeatureKey,
+} from './student/store/student.reducer';
+
+const REDUCERS = {
+  [teachersFeatureKey]: teacherReducer,
+  [studentsFeatureKey]: studentReducer,
+};
+
 export const appConfig: ApplicationConfig = {
   providers: [
     provideStore(REDUCERS),
